lab7/requirements.txt
asgiref==3.7.2
blinker==1.4
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
cmdtest==0.32+git
command-not-found==0.3
cryptography==3.4.8
db.py==0.5.3
dbus-python==1.2.18
devscripts===2.22.1ubuntu1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.2.2
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
filelock==3.6.0
fontawesomefree==6.4.0
gpg===1.16.0-unknown
greenlet==2.0.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
more-itertools==8.10.0
msgpack==1.0.4
mutagen==1.45.1
mysql-connector-python==8.0.33
netifaces==0.11.0
numpy==1.25.0
oauthlib==3.2.0
pandas==2.0.2
pbr==5.11.1
Pillow==9.5.0
platformdirs==2.5.1
prettytable==0.7.2
protobuf==3.20.3
pybars3==0.9.7
pycryptodomex==3.11.0
pygame==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
PyMeta3==0.5.1
pynvim==0.4.3
pyparsing==2.4.7
python-apt==2.4.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-magic==0.4.24
pytz==2023.3
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
stevedore==5.1.0
systemd-python==234
ttystatus==0.38
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
unidiff==0.5.5
urllib3==1.26.5
virtualenv==20.13.0+ds
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.6
websockets==9.1
xdg==5
yt-dlp==2022.4.8
zipp==1.0.0
